fix(graphrag): Improve robustness and error handling in GraphRAG modules
Ensure Qdrant client is always initialized by switching to `get_qdrant_client_required()` in `ArchitectureMapper`, `CodeNavigator`, and `DependencyTracer`. This prevents potential runtime errors if the client is not configured.

Enhance cycle detection in `DependencyTracer` by explicitly skipping cycles longer than 10 nodes and adding null checks when retrieving entity names to avoid errors from missing entities. Also added a type ignore for `nx.simple_cycles` to address type checking.

Improve type safety in `InteractionTracer` by using `typing.cast` for the `dspy.Prediction` return type.

Provide more informative error messages in `TemporalExtractor` when URLs resolve to unsafe or unresolvable IP addresses, using specific error details from `_get_safe_ip` for better user feedback.

Update `uv.lock` due to these changes.

+++ server/infrastructure/qdrant/async_wrapper.py
@@ -28,7 +28,7 @@ class QdrantAsyncWrapper:
         await wrapper.async_upsert(...)  # Non-blocking
     """
 
-    def __init__(self, url: str, prefer_async: bool = False, timeout: float = 90.0):
+    def __init__(self, url: str, prefer_async: bool = False, timeout: int = 90):
         """
         Initialize dual-mode client.
